Handle combined download and extraction stages
resin-os/docker extracts the layer as it’s downloaded resulting in progress reporting stopping at 50% (since there’s no “Downloading” phase anymore).

We determine that’s the case on-the-fly, by looking for the “Ready to download” event status, in which case we create a “link” from the download progress tracker to the extraction tracker for that layer by sharing the object reference. This let’s us transparently handle the coalescing of the two stages and still report an accurate percentage.

Change-Type: minor
Connects-To: #32
Refs: https://github.com/resin-os/balaena/blob/f55283222c05345540a0654cea2ef15031909806/distribution/pull_v2.go#L230