pull: fix digest hash parsing
The current code assumes the digest status line is in the form
`Digest: <alg>:<hash>`, taking up the whole line. This is not at all the
case on all docker daemon versions, which is now in the form
`<tag>: digest: <alg>:<hash> size: <size>`.

Fix the regex to accept both formats:

- accept leading characters before the digest part
- make the `d` case-insensitive
- accept trailing elements after the hash

Change-type: patch